+++ index.js
@@ -18,6 +18,7 @@ var exchangeMetadata = require('./lib/exchange-metadata');
 var storage = require('./lib/storage');
 var fileStream = require('./lib/file-stream');
 var piece = require('./lib/piece');
+var rechoke = require('./lib/rechoke');
 
 var MAX_REQUESTS = 5;
 var CHOKE_TIMEOUT = 5000;
@@ -29,7 +30,6 @@ var BAD_PIECE_STRIKES_MAX = 3;
 var BAD_PIECE_STRIKES_DURATION = 120000; // 2 minutes
 
 var RECHOKE_INTERVAL = 10000;
-var RECHOKE_OPTIMISTIC_DURATION = 2;
 
 var TMP = fs.existsSync('/tmp') ? '/tmp' : os.tmpDir();
 
@@ -93,8 +93,6 @@ var torrentStream = function(link, opts, cb) {
 	var refresh = noop;
 
 	var rechokeSlots = (opts.uploads === false || opts.uploads === 0) ? 0 : (+opts.uploads || 10);
-	var rechokeOptimistic = null;
-	var rechokeOptimisticTime = 0;
 	var rechokeIntervalId;
 
 	engine.infoHash = infoHash;
@@ -458,66 +456,7 @@ var torrentStream = function(link, opts, cb) {
 			id = setTimeout(onchoketimeout, timeout);
 		};
 
-		var rechokeSort = function(a, b) {
-			// Prefer higher download speed
-			if (a.downSpeed !== b.downSpeed) return a.downSpeed > b.downSpeed ? -1 : 1;
-			// Prefer higher upload speed
-			if (a.upSpeed !== b.upSpeed) return a.upSpeed > b.upSpeed ? -1 : 1;
-			// Prefer unchoked
-			if (a.wasChoked !== b.wasChoked) return a.wasChoked ? 1 : -1;
-			// Random order
-			return a.salt - b.salt;
-		};
-
-		var onrechoke = function() {
-			if (rechokeOptimisticTime > 0) --rechokeOptimisticTime;
-			else rechokeOptimistic = null;
-
-			var peers = [];
-
-			wires.forEach(function(wire) {
-				if (wire.isSeeder) {
-					if (!wire.amChoking) wire.choke();
-				} else if (wire !== rechokeOptimistic) {
-					peers.push({
-						wire:       wire,
-						downSpeed:  wire.downloadSpeed(),
-						upSpeed:    wire.uploadSpeed(),
-						salt:       Math.random(),
-						interested: wire.peerInterested,
-						wasChoked:  wire.amChoking,
-						isChoked:   true
-					});
-				}
-			});
-
-			peers.sort(rechokeSort);
-
-			var i = 0;
-			var unchokeInterested = 0;
-			for (; i < peers.length && unchokeInterested < rechokeSlots; ++i) {
-				peers[i].isChoked = false;
-				if (peers[i].interested) ++unchokeInterested;
-			}
-
-			if (!rechokeOptimistic && i < peers.length && rechokeSlots) {
-				var candidates = peers.slice(i).filter(function(peer) { return peer.interested; });
-				var optimistic = candidates[(Math.random() * candidates.length) | 0];
-
-				if (optimistic) {
-					optimistic.isChoked = false;
-					rechokeOptimistic = optimistic.wire;
-					rechokeOptimisticTime = RECHOKE_OPTIMISTIC_DURATION;
-				}
-			}
-
-			peers.forEach(function(peer) {
-				if (peer.wasChoked !== peer.isChoked) {
-					if (peer.isChoked) peer.wire.choke();
-					else peer.wire.unchoke();
-				}
-			});
-		};
+		var onrechoke = rechoke(wires, rechokeSlots);
 
 		var onready = function() {
 			swarm.on('wire', onwire);
